+++ CHANGES.md
@@ -23,6 +23,9 @@
 
 - Added `Sys.argv`.
 
+- Deprecated `Not_found`, people who need it can use `Caml.Not_found`, but its
+  use isn't recommended.
+
 ## v0.10
 
 (Changes that can break existing programs are marked with a "\*")